server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
tweepy==4.14.0
requests
beautifulsoup4
